Namespaces b/c this is strictly relevant to EVSS
